refactor(coverde-cli): relocate custom instances
