+++ packages/coverde_cli/lib/src/commands/filter/filter.dart
@@ -1,4 +1,4 @@
-import 'package:coverde/coverde.dart';
+import 'package:coverde/src/commands/coverde_command.dart';
 import 'package:coverde/src/entities/trace_file.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
